Further reduce address parsing
